+++ Viewer/src/VLabelAttr.cpp
@@ -130,8 +130,8 @@ void VLabelAttr::scan(VNode* vnode,std::vector<VAttribute*>& vec)
     if(vnode->node_)
     {
         const std::vector<Label>& v=vnode->node_->labels();
-        int n=v.size();
-        for(size_t i=0; i < n; i++)
+        int n=static_cast<int>(v.size());
+        for(int i=0; i < n; i++)
         {
             vec.push_back(new VLabelAttr(vnode,v[i],i));
         }
