+++ Viewer/src/VUserVarAttr.cpp
@@ -111,8 +111,8 @@ void VUserVarAttr::scan(VNode* vnode,std::vector<VAttribute*>& vec)
         if(vnode->node_)
         {
             const std::vector<Variable>& v=vnode->node_->variables();
-            int n=v.size();
-            for(size_t i=0; i < n; i++)
+            int n=static_cast<int>(v.size());
+            for(int i=0; i < n; i++)
             {
                 vec.push_back(new VUserVarAttr(vnode,v[i],i));
             }
@@ -123,8 +123,8 @@ void VUserVarAttr::scan(VNode* vnode,std::vector<VAttribute*>& vec)
     {
         std::vector<Variable> v;
         vnode->variables(v);
-        int n=v.size();
-        for(size_t i=0; i < n; i++)
+        int n=static_cast<int>(v.size());
+        for(int i=0; i < n; i++)
         {
             vec.push_back(new VUserVarAttr(vnode,v[i],i));
         }
